split matchmaking into 3 files
