+++ hivemind/client/averaging/matchmaking.py
@@ -2,21 +2,22 @@
 
 from __future__ import annotations
 
-import contextlib
 import random
 from dataclasses import asdict
 from math import isfinite
-from typing import Sequence, Optional, AsyncIterator, Set, Tuple, Dict
+from typing import Sequence, Optional, AsyncIterator, Dict
 import asyncio
 
 import grpc
 import torch
 
-import hivemind
 from hivemind.client.averaging.allreduce import AllReduceRunner
+from hivemind.client.averaging.dht_handler import DHTHandler
+from hivemind.client.averaging.leaders_queue import PotentialLeaders
 from hivemind.client.averaging.load_balancing import load_balance_peers
-from hivemind.dht import DHTID, DHTExpiration, get_dht_time, GroupKey
-from hivemind.utils import get_logger, Endpoint, TensorDescriptor, MSGPackSerializer, TimedStorage
+from hivemind.client.averaging.group_key import GroupKeyManager
+from hivemind.dht import DHT, DHTID, DHTExpiration, get_dht_time
+from hivemind.utils import get_logger, Endpoint, TensorDescriptor, MSGPackSerializer
 from hivemind.proto import averaging_pb2, averaging_pb2_grpc
 from hivemind.utils.grpc import ChannelCache
 
@@ -36,7 +37,7 @@ class Matchmaking(averaging_pb2_grpc.DecentralizedAveragingServicer):
       Hence, instead of accounting for such deadlocks, we simply break them with request_timeout.
     """
 
-    def __init__(self, endpoint: Endpoint, averaged_tensors: Sequence[torch.Tensor], dht: hivemind.dht.DHT, *,
+    def __init__(self, endpoint: Endpoint, averaged_tensors: Sequence[torch.Tensor], dht: DHT, *,
                  prefix: str, target_group_size: int, min_group_size: int, initial_group_bits: Optional[str] = None,
                  averaging_expiration: float = 15, request_timeout: float, throughput: Optional[float] = None,
                  min_vector_size: int, **allreduce_kwargs):
@@ -46,12 +47,11 @@ class Matchmaking(averaging_pb2_grpc.DecentralizedAveragingServicer):
                            "matchmaking can cause deadlocks in some rare cases. Please see Matchmaking docstring.")
 
         super().__init__()
-        self.dht, self.endpoint, self.averaged_tensors = dht, endpoint, tuple(averaged_tensors)
-        self.prefix, self.group_bits = prefix, initial_group_bits
+        self.endpoint, self.averaged_tensors, self.dht_handler = endpoint, tuple(averaged_tensors), DHTHandler(dht)
+        self.group_key_manager = GroupKeyManager(self.dht_handler, prefix, initial_group_bits)
         self.target_group_size, self.min_group_size = target_group_size, min_group_size
         self.averaging_expiration, self.request_timeout = averaging_expiration, request_timeout
-        self.throughput, self.min_vector_size = throughput, min_vector_size
-        self.allreduce_kwargs = allreduce_kwargs
+        self.throughput, self.min_vector_size, self.allreduce_kwargs = throughput, min_vector_size, allreduce_kwargs
         self.schema_hash = compute_schema_hash(self.averaged_tensors)
         self.total_size = sum(tensor.numel() for tensor in self.averaged_tensors)
 
@@ -63,16 +63,12 @@ class Matchmaking(averaging_pb2_grpc.DecentralizedAveragingServicer):
 
         self.current_leader: Optional[Endpoint] = None  # iff i am a follower, this is a link to my current leader
         self.current_followers: Dict[Endpoint, averaging_pb2.JoinRequest] = {}  # my current followers excluding myself
-        self.potential_leaders = PotentialLeaders(endpoint, dht, averaging_expiration, target_group_size)
-        self.data_for_gather: bytes = None
+        self.potential_leaders = PotentialLeaders(endpoint, self.dht_handler, averaging_expiration, target_group_size)
+        self.data_for_gather: Optional[bytes] = None
 
     @property
     def is_looking_for_group(self):
         return self.lock_looking_for_group.locked()
-
-    @property
-    def current_group_key(self) -> GroupKey:
-        return f"{self.prefix}.0b{self.group_bits}"
 
     def __repr__(self):
         lfg_status = "looking for group," if self.is_looking_for_group else "not looking for group,"
@@ -83,12 +79,12 @@ class Matchmaking(averaging_pb2_grpc.DecentralizedAveragingServicer):
                 lfg_status += f" leading {len(self.current_followers)} followers,"
         schema_hash_repr = f"{self.schema_hash[0]}...{self.schema_hash[-8:]}"
         return f"{self.__class__.__name__}(endpoint={self.endpoint}, schema={schema_hash_repr}, {lfg_status}" \
-               f" current key = {self.current_group_key})"
+               f" current key = {self.group_key_manager.current_key})"
 
     async def look_for_group(self, *, data_for_gather: bytes = b'', timeout: Optional[float] = None
                              ) -> Optional[AllReduceRunner]:
         """
-        :param gather: optionally send this data to all peers in the next group and gather it from every groupmate
+        :param data_for_gather: optionally send this data to all peers in the next group and gather it from every groupmate
         :param timeout: maximum time that may be spent looking for group (does not include allreduce itself)
         :returns: an assembled group if successful, None if failed; does NOT perform the actual averaging
         Iterate over the averagers from a given group_identifier that have higher leadership priority than yourself.
@@ -127,10 +123,7 @@ class Matchmaking(averaging_pb2_grpc.DecentralizedAveragingServicer):
 
     async def _request_join_potential_leaders(self, timeout: Optional[float]) -> AllReduceRunner:
         """ Request leaders from queue until we find the first runner. This coroutine is meant to run in background. """
-        async with self.potential_leaders.begin_search(self.current_group_key, timeout):
-            # TODO update group_bits on success! reduce number of bits on not enough peers.
-            # TODO after allreduce finishes, we may need to ask leader to notify lower keys about this
-            # (so as to fix possible network partitioning if some peers operate on a much smaller nbits)
+        async with self.potential_leaders.begin_search(self.group_key_manager.current_key, timeout):
             while True:
                 try:
                     next_leader = await self.potential_leaders.pop_next_leader()  # throws TimeoutError on expiration
@@ -147,8 +140,7 @@ class Matchmaking(averaging_pb2_grpc.DecentralizedAveragingServicer):
                             # the time is up, we have a *good enough* group. run allreduce as is.
                             return await self.leader_assemble_group()
                         elif len(self.current_followers) > 0:
-                            await self.leader_disband_group()
-                            # TODO maybe adjust grid size
+                            await self.leader_disband_group(update_key=True)
                         continue
                 except Exception as e:
                     if not self.assembled_group.done():
@@ -246,7 +238,7 @@ class Matchmaking(averaging_pb2_grpc.DecentralizedAveragingServicer):
                         # outcome 2: the time is up, run allreduce with what we have or disband
                         await self.leader_assemble_group()
                     else:
-                        await self.leader_disband_group()
+                        await self.leader_disband_group(update_key=True)
 
             if self.was_accepted_to_group.is_set() or not self.assembled_group.done() \
                     or self.assembled_group.cancelled() or request.endpoint not in self.assembled_group.result():
@@ -262,8 +254,8 @@ class Matchmaking(averaging_pb2_grpc.DecentralizedAveragingServicer):
             allreduce_group = self.assembled_group.result()
             yield averaging_pb2.MessageFromLeader(
                 code=averaging_pb2.BEGIN_ALLREDUCE, group_id=allreduce_group.group_id,
-                ordered_group_endpoints=allreduce_group.ordered_group_endpoints,
-                part_sizes=allreduce_group.part_sizes, gathered=allreduce_group.gathered)
+                ordered_group_endpoints=allreduce_group.ordered_group_endpoints, part_sizes=allreduce_group.part_sizes,
+                gathered=allreduce_group.gathered, group_key_seed=allreduce_group.group_key_seed)
 
         except Exception as e:
             logger.exception(e)
@@ -319,11 +311,13 @@ class Matchmaking(averaging_pb2_grpc.DecentralizedAveragingServicer):
                 gathered.append(follower_info.gather if follower_info.gather else None)
 
         part_sizes = load_balance_peers(self.total_size, throughputs, self.min_vector_size)
+        group_key_seed = random.randint(- 2 ** 31, 2 ** 31 - 1)
 
         logger.debug(f"{self.endpoint} - leader started allreduce for {len(ordered_group_endpoints)} peers.")
         allreduce_group = AllReduceRunner(group_id=group_id, tensors=self.averaged_tensors, endpoint=self.endpoint,
                                           ordered_group_endpoints=ordered_group_endpoints, part_sizes=part_sizes,
-                                          gathered=gathered, **self.allreduce_kwargs)
+                                          gathered=gathered, group_key_seed=group_key_seed, **self.allreduce_kwargs)
+        self.group_key_manager.update_key_on_success(allreduce_group)
         self.assembled_group.set_result(allreduce_group)
         return allreduce_group
 
@@ -340,136 +334,18 @@ class Matchmaking(averaging_pb2_grpc.DecentralizedAveragingServicer):
         logger.debug(f"{self.endpoint} - follower started allreduce after being prompted by leader {leader}.")
         allreduce_group = AllReduceRunner(group_id=group_id, tensors=self.averaged_tensors, endpoint=self.endpoint,
                                           ordered_group_endpoints=tuple(ordered_group_endpoints),
-                                          part_sizes=tuple(part_sizes), gathered=msg.gathered, **self.allreduce_kwargs)
+                                          part_sizes=tuple(part_sizes), gathered=msg.gathered,
+                                          group_key_seed=int(msg.group_key_seed), **self.allreduce_kwargs)
+        self.group_key_manager.update_key_on_success(allreduce_group)
         self.assembled_group.set_result(allreduce_group)
         return allreduce_group
 
-    async def leader_disband_group(self):
+    async def leader_disband_group(self, update_key=False):
         """ Kick out all followers immediately, optionally direct them to our new leader (if we found one) """
         assert self.lock_request_join_group.locked()
         self.current_followers.clear()  # this will cause rpc_join_group to kick all followers out
-
-
-class PotentialLeaders:
-    """ An utility class that searches for averagers that could become our leaders """
-
-    def __init__(self, endpoint: Endpoint, dht: hivemind.DHT, averaging_expiration: DHTExpiration,
-                 target_group_size: Optional[int]):
-        self.endpoint, self.dht, self.averaging_expiration = endpoint, dht, averaging_expiration
-        self.target_group_size = target_group_size
-        self.running, self.update_triggered, self.update_finished = asyncio.Event(), asyncio.Event(), asyncio.Event()
-        self.declared_expiration, self.lock_search, self.lock_declare = asyncio.Event(), asyncio.Lock(), asyncio.Lock()
-        self.leader_queue = TimedStorage[Endpoint, DHTExpiration]()
-        self.past_attempts: Set[Tuple[Endpoint, DHTExpiration]] = set()
-        self.declared_expiration_time = float('inf')
-        self.declared_group_key: Optional[GroupKey] = None
-        self.max_assured_time = float('-inf')
-        self.search_end_time = float('inf')
-
-    @contextlib.asynccontextmanager
-    async def begin_search(self, group_key: GroupKey, timeout: Optional[float]):
-        async with self.lock_search:
-            self.running.set()
-            self.search_end_time = get_dht_time() + timeout if timeout is not None else float('inf')
-            update_queue_task = asyncio.create_task(self._update_queue_periodically(group_key))
-            declare_averager_task = asyncio.create_task(self._declare_averager_periodically(group_key))
-            try:
-                yield self
-            finally:
-                if not update_queue_task.done():
-                    update_queue_task.cancel()
-                if not declare_averager_task.done():
-                    declare_averager_task.cancel()
-                for field in (self.past_attempts, self.leader_queue, self.running,
-                              self.update_finished, self.update_triggered, self.declared_expiration):
-                    field.clear()
-                self.max_assured_time = float('-inf')
-                self.search_end_time = float('inf')
-
-    @contextlib.asynccontextmanager
-    async def pause_search(self):
-        was_running = self.running.is_set()
-        try:
-            self.running.clear()
-            yield
-        finally:
-            if was_running:
-                self.running.set()
-            else:
-                self.running.clear()
-
-    async def pop_next_leader(self) -> Endpoint:
-        """ Remove and return the next most suitable leader or throw an exception if reached timeout """
-        assert self.running.is_set(), "Not running search at the moment"
-        while True:
-            maybe_next_leader, entry = self.leader_queue.top()
-
-            if maybe_next_leader is None or self.max_assured_time <= entry.expiration_time <= self.search_end_time:
-                self.update_triggered.set()
-
-            if maybe_next_leader is None or entry.expiration_time >= self.declared_expiration_time:
-                await asyncio.wait({self.update_finished.wait(), self.declared_expiration.wait()},
-                                   return_when=asyncio.FIRST_COMPLETED)
-                self.declared_expiration.clear()
-                if self.update_finished.is_set():
-                    self.update_finished.clear()
-                    continue
-                else:
-                    raise asyncio.TimeoutError("pop_next_leader was invalidated: re-declared averager in background")
-
-            del self.leader_queue[maybe_next_leader]
-            self.past_attempts.add((maybe_next_leader, entry.expiration_time))
-            return maybe_next_leader
-
-    @property
-    def request_expiration_time(self) -> float:
-        """ this averager's current expiration time - used to send join requests to leaders """
-        if isfinite(self.declared_expiration_time):
-            return self.declared_expiration_time
-        else:
-            return min(get_dht_time() + self.averaging_expiration, self.search_end_time)
-
-    async def _update_queue_periodically(self, group_key: GroupKey):
-        DISCREPANCY = hivemind.utils.timed_storage.MAX_DHT_TIME_DISCREPANCY_SECONDS
-        while get_dht_time() < self.search_end_time:
-            new_peers = await self.dht.get_averagers(group_key, only_active=True, return_future=True)
-            self.max_assured_time = max(self.max_assured_time, get_dht_time() + self.averaging_expiration - DISCREPANCY)
-
-            self.leader_queue.clear()
-            for peer, peer_expiration_time in new_peers:
-                if peer == self.endpoint or (peer, peer_expiration_time) in self.past_attempts:
-                    continue
-                self.leader_queue.store(peer, peer_expiration_time, peer_expiration_time)
-                self.max_assured_time = max(self.max_assured_time, peer_expiration_time - DISCREPANCY)
-
-            self.update_finished.set()
-
-            await asyncio.wait(
-                {self.running.wait(), self.update_triggered.wait()}, return_when=asyncio.ALL_COMPLETED,
-                timeout=self.search_end_time - get_dht_time() if isfinite(self.search_end_time) else None)
-            self.update_triggered.clear()
-
-    async def _declare_averager_periodically(self, group_key: GroupKey):
-        async with self.lock_declare:
-            try:
-                while True:
-                    await self.running.wait()
-
-                    new_expiration_time = min(get_dht_time() + self.averaging_expiration, self.search_end_time)
-                    self.declared_group_key, self.declared_expiration_time = group_key, new_expiration_time
-                    self.declared_expiration.set()
-                    await self.dht.declare_averager(group_key, self.endpoint, new_expiration_time,
-                                                    looking_for_group=True, return_future=True)
-                    await asyncio.sleep(self.declared_expiration_time - get_dht_time())
-            except Exception as e:  # note: we catch exceptions here because otherwise they are never printed
-                logger.error(f"{self.endpoint} - caught {type(e)}: {e}")
-            finally:
-                if self.declared_group_key is not None:
-                    prev_declared_key, prev_expiration_time = self.declared_group_key, self.declared_expiration_time
-                    self.declared_group_key, self.declared_expiration_time = None, float('inf')
-                    self.leader_queue, self.max_assured_time = TimedStorage[Endpoint, DHTExpiration](), float('-inf')
-                    await self.dht.declare_averager(prev_declared_key, self.endpoint, prev_expiration_time,
-                                                    looking_for_group=False, return_future=True)
+        if update_key:
+            self.group_key_manager.update_key_on_failure()
 
 
 def compute_schema_hash(tensors: Sequence[torch.Tensor]) -> bytes:
