+++ tests/test_averaging.py
@@ -7,26 +7,28 @@ import pytest
 import hivemind
 from hivemind.client.averaging.allreduce import AllReduceProtocol, split_into_parts, restore_from_parts
 from hivemind.client.averaging.load_balancing import load_balance_peers
+from hivemind.client.averaging.dht_handler import DHTHandler
 from hivemind.utils import Endpoint
 
 
 @pytest.mark.forked
-def test_getset_averagers():
-    dht = hivemind.DHT(start=True)
+@pytest.mark.asyncio
+async def test_getset_averagers():
+    dht_handler = DHTHandler(hivemind.DHT(start=True))
 
     t = hivemind.get_dht_time()
-    dht.declare_averager(group_key='bucket.0b10110', endpoint='localhvost', expiration_time=t + 60)
-    dht.declare_averager(group_key='bucket.0b10110', endpoint='localhvost2', expiration_time=t + 61)
+    await dht_handler.declare_averager(group_key='bucket.0b10110', endpoint='localhvost', expiration_time=t + 60)
+    await dht_handler.declare_averager(group_key='bucket.0b10110', endpoint='localhvost2', expiration_time=t + 61)
 
-    q1 = dht.get_averagers('bucket.0b10110', only_active=True)
+    q1 = await dht_handler.get_averagers('bucket.0b10110', only_active=True)
 
-    dht.declare_averager(group_key='bucket.0b10110', endpoint='localhvost', expiration_time=t + 66)
-    q2 = dht.get_averagers('bucket.0b10110', only_active=True)
+    await dht_handler.declare_averager(group_key='bucket.0b10110', endpoint='localhvost', expiration_time=t + 66)
+    q2 = await dht_handler.get_averagers('bucket.0b10110', only_active=True)
 
-    dht.declare_averager(group_key='bucket.0b10110', endpoint='localhvost2', looking_for_group=False,
-                         expiration_time=t + 61)
-    q3 = dht.get_averagers('bucket.0b10110', only_active=True)
-    q4 = dht.get_averagers('bucket.0b10110', only_active=False)
+    await dht_handler.declare_averager(group_key='bucket.0b10110', endpoint='localhvost2', looking_for_group=False,
+                                       expiration_time=t + 61)
+    q3 = await dht_handler.get_averagers('bucket.0b10110', only_active=True)
+    q4 = await dht_handler.get_averagers('bucket.0b10110', only_active=False)
 
     assert len(q1) == 2 and ('localhvost', t + 60) in q1 and ('localhvost2', t + 61) in q1
     assert len(q2) == 2 and ('localhvost', t + 66) in q2 and ('localhvost2', t + 61) in q2
